requirements.txt
requests==2.32.3
datetime==5.5
python-dotenv==1.0.1